requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.0
django-crispy-forms==1.9.0
django-filter==2.2.0
django-formtools==2.1
django-guardian==2.1.0
django-import-export==2.1.0
django-reversion==3.0.7
djangorestframework==3.11.0
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
idna==2.9
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
Markdown==3.2.1
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==1.4.6
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
six==1.10.0
sqlparse==0.3.1
tablib==1.1.0
uritemplate==3.0.1
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
